fix(storybook): merge e2e-storybook-angular into e2e-storybook
