+++ .github/workflows/e2e-windows.yml
@@ -69,7 +69,6 @@ jobs:
           - e2e-web
           - e2e-rollup
           - e2e-storybook
-          - e2e-storybook-angular
           - e2e-vite
           - e2e-webpack
           - e2e-workspace-create
@@ -118,8 +117,6 @@ jobs:
           - project: e2e-rollup
             codeowners: 'S04SJ6PL98X'
           - project: e2e-storybook
-            codeowners: 'S04SVQ8H0G5'
-          - project: e2e-storybook-angular
             codeowners: 'S04SVQ8H0G5'
           - project: e2e-vite
             codeowners: 'S04SJ6PL98X'
