--- e2e/storybook-angular/project.json
@@ -1,11 +1,0 @@
-{
-  "name": "e2e-storybook-angular",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "e2e/storybook-angular",
-  "projectType": "application",
-  "targets": {
-    "e2e": {},
-    "run-e2e-tests": {}
-  },
-  "implicitDependencies": ["storybook"]
-}
